feat:provide rank stat

+++ internal/model/piece/response.go
@@ -30,7 +30,7 @@ type FacultyStats struct {
 type CollectedPiecesStats struct {
 	TotalCollected     int                     `json:"total_collected"`
 	CollectedByFaculty map[string]FacultyStats `json:"collected_by_faculty"`
-	SameMissingCount   map[int]float64         `json:"same_missing_count"`
+	SameMissingCount   map[int]int             `json:"same_missing_count"`
 	Rank               int                     `json:"rank"`
 }
